Audio transitions fixed & other edits:
Audio manager was unhappy if you tried to tell it to transition to new music while it was still in the process of fading in music, that should be fixed now.
Music restarting when going between options or controls back to title page should be fixed.
Issue where the loading screen doesn't go away but the game is playing might be fixed by just making sure the loading screen turns off once a new scene starts.
Was getting a minor error in the credits about some animation-event-message-function-something wasn't defined. Found two empty animation events in the credits animation and went ahead and deleted them.