fixed bug making webgl not work on some browsers
There was an infinite loading bug on some browsers, this should fix it but makes refreshing  the page kill saves. Makes the game session based instead of system based which sucks but the way to fix it is making it only play on chrome or completely making a new save system and overall game rework to fit that.